nl80211: Fix backport of struct wiphy_vendor_command
Commit 21b770cb4a6e4278c2b3cbe96a97641238bb7d66 ("nl80211: vendor
command support") was used to backport nl80211 vendor command support
from 3.14.  Unfortunately this backport has a typo in the definition
of struct wiphy_vendor_command, namely the third param to the doit
callback function was declared as "void *data" instead of the correct
"const void *data".  Correct this flaw in the backport.

Change-Id: I1caf72215ebea94f6f9e7a930ab54d67ec8c2538
CRs-Fixed: 768724
Signed-off-by: Jeff Johnson <jjohnson@codeaurora.org>
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>