msm: kgsl: Wake up snoozing threads on marker expiry
Marker commands may not be submitted to RB but
create room in context queue on expiry which is
left uncounted. If there are too many markers
commands from a specific context then context
queue eventually gets full but wakeup is never
called for sleeping threads.

Wake up the snoozing threads when there is room
in the context queue rather than based on the
condition of real submissions from the context.

CRs-fixed: 783187

Change-Id: I0cb0002d7b641071a45d46ae50ce9f23f37b0fab
Signed-off-by: Sunil Khatri <sunilkh@codeaurora.org>
Signed-off-by: Anupam Sakargayan <asakarga@codeaurora.org>
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>